Updated BlockHeader struct

+++ src/chainweb_client.rs
@@ -1,3 +1,4 @@
+use bigdecimal::BigDecimal;
 use futures::{Stream, TryStreamExt};
 use reqwest::Url;
 use serde::Deserializer;
@@ -64,7 +65,7 @@ pub struct BlockHeader {
     pub payload_hash: String,
     pub weight: String,
     #[serde(rename(deserialize = "featureFlags"))]
-    pub feature_flags: i64,
+    pub feature_flags: BigDecimal,
     #[serde(rename(deserialize = "epochStart"))]
     pub epoch_start: i64,
     pub adjacents: HashMap<ChainId, String>,
